+++ app/src/main/java/com/philipgreen/dmwizard/player/utils/BaseStats.java
@@ -1,4 +1,4 @@
-package com.philipgreen.dmwizard.data;
+package com.philipgreen.dmwizard.player.utils;
 
 /**
  * Created by pgreen on 8/6/16.
