+++ app/src/main/java/com/philipgreen/dmwizard/playerClasses/BasePlayerClass.java
@@ -1,7 +1,7 @@
 package com.philipgreen.dmwizard.playerClasses;
 
 import com.philipgreen.dmwizard.data.ArmorProficiencies;
-import com.philipgreen.dmwizard.data.BaseStats;
+import com.philipgreen.dmwizard.player.utils.BaseStats;
 import com.philipgreen.dmwizard.player.utils.Skills;
 import com.philipgreen.dmwizard.data.WeaponProficiencies;
